Organized basic architecture in frontend
