--- package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "cypress": "^10.8.0"
-  }
-}
