+++ web-ui/src/App.js
@@ -5,19 +5,23 @@ import {
     Routes,
     Navigate
 } from 'react-router-dom'
-import HomePageForm from './HomePageForm'
-import AddHealthReport from './AddHealthReport'
-import CrearActivoFijoForm from './components/CrearActivoFijoForm'
-import ShowFilesForm from './ShowFilesForm'
-import LoginForm from './components/LoginForm'
-import ShowFixedAssets from './components/ShowFixedAssets'
-import DataHealth from './kidsFilesView/DataHealth'
-import { ShowFixedAsset } from './components/ShowOneFixedAsset'
+import HomePageForm from './Views/HomePage/HomePageForm'
+import LoginForm from './Views/Login/LoginForm'
+
+import CrearActivoFijoForm from './Views/FixedAssets/CrearActivoFijoForm'
+import ShowFixedAssets from './Views/FixedAssets/ShowFixedAssets'
+import { ShowFixedAsset } from './Views/FixedAssets/ShowOneFixedAsset'
+
+import AddHealthReport from './Views/KidsFiles/AddHealthReport'
+import ShowFilesForm from './Views/KidsFiles/ShowFilesForm'
+import DataHealth from './Views/KidsFiles/DataHealth'
 
 function App() {
     return (
         <Router>
             <Routes>
+                <Route exact path="/" element={<LoginForm />}></Route>
+                <Route path="/home-ncv" element={<HomePageForm />}></Route>
                 <Route
                     path="/crear-activo-fijo"
                     element={<CrearActivoFijoForm />}
@@ -26,17 +30,16 @@ function App() {
                     path="/activos-fijos"
                     element={<ShowFixedAssets />}
                 ></Route>
+                
                 <Route
                     path="/activos-fijos/:fixedAssetId"
                     element={<ShowFixedAsset />}
                 ></Route>
                 <Route path="/files-nenes" element={<ShowFilesForm />}></Route>
-                <Route path="/home-ncv" element={<HomePageForm />}></Route>
                 <Route
                     path="/add-reporte-nene"
                     element={<AddHealthReport />}
                 ></Route>
-                <Route exact path="/" element={<LoginForm />}></Route>
                 <Route path="kidHealth" element={<DataHealth />} />
                 <Route path="*" element={<Navigate replace to="/" />} />
             </Routes>
