+++ web-ui/src/Views/FixedAssets/ShowFixedAssets.js
