+++ web-ui/src/index.js
@@ -1,8 +1,5 @@
 // import React from 'react';
 import ReactDOM from 'react-dom/client'
-import HomePageForm from './HomePageForm'
-import LoginForm from './components/LoginForm'
-import CrearActivoFijoForm from './components/CrearActivoFijoForm'
 import App from './App'
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
